Consolidate existing clients into just 2 clients: "Token" and "Transaction"
