+++ Client/src/main/java/com/seamlesspay/api/Transaction.java
@@ -8,10 +8,13 @@
 package com.seamlesspay.api;
 
 import com.seamlesspay.api.interfaces.BaseChargeTokenCallback;
+import com.seamlesspay.api.interfaces.RefundTokenCallback;
 import com.seamlesspay.api.models.BaseChargeToken;
 import com.seamlesspay.api.models.CardChargeBulder;
+import com.seamlesspay.api.models.RefundBuilder;
+import com.seamlesspay.api.models.RefundToken;
 
-public class Charge {
+public class Transaction {
 
   public static void create(
     final SeamlesspayFragment fragment,
@@ -32,6 +35,28 @@ public class Charge {
           fragment.postCallback(exception);
         }
       }
+    );
+  }
+
+  public static void create(
+      final SeamlesspayFragment fragment,
+      final RefundBuilder refundBuilder
+  ) {
+    RefundClient.create(
+        fragment,
+        refundBuilder,
+        new RefundTokenCallback() {
+
+          @Override
+          public void success(RefundToken refundToken) {
+            fragment.postCallback(refundToken);
+          }
+
+          @Override
+          public void failure(Exception exception) {
+            fragment.postCallback(exception);
+          }
+        }
     );
   }
 }
